refactor(tools): hoist shared mmgisAPI adapters into Tools/_shared
mmgisAPI.ts and useMMGISHandlerReady.ts were byte-identical in both Card and
LayerManager. Move a single copy to src/essence/Tools/_shared/adapters/,
delete the duplicates, and repoint imports in both tools so future tools can
share one source of truth. No behavior change; tsc --noEmit passes.
