+++ src/essence/Tools/Card/CardTool.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { createRoot, type Root } from 'react-dom/client'
 import { MMGISCardAdapter } from './MMGISCardAdapter'
-import { mmgisRequest } from './adapters/mmgisAPI'
+import { mmgisRequest } from '../_shared/adapters/mmgisAPI'
 
 type ToolVars = { width?: number }
