+++ src/essence/Tools/LayerManager/LayerManagerTool.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import $ from 'jquery'
 import { createRoot, type Root } from 'react-dom/client'
 import { MMGISLayerManagerAdapter } from './MMGISLayerManagerAdapter'
-import { mmgisRequest, mmgisProvide, mmgisEmit } from './adapters/mmgisAPI'
+import { mmgisRequest, mmgisProvide, mmgisEmit } from '../_shared/adapters/mmgisAPI'
 
 type ToolVars = { width?: number }
