+++ src/essence/Tools/LayerManager/adapters/getVisibleLayersWithLegends.ts
@@ -1,4 +1,4 @@
-import { mmgisRequest } from './mmgisAPI'
+import { mmgisRequest } from '../../_shared/adapters/mmgisAPI'
 import { buildLayerLegendData } from './buildLayerLegendData'
 import type { Layer } from '../lib/types'
