+++ src/essence/Tools/_shared/adapters/useMMGISHandlerReady.ts
